Satisfy Rust file size check
