+++ docs/case-studies/issue-20/evidence/cargo-clippy.log
@@ -1,2 +1,1 @@
-    Checking link-cli v0.1.0 (/tmp/gh-issue-solver-1778236390702/rust)
-    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.80s
+    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.10s
