+++ docs/case-studies/issue-20/evidence/rust-file-size-check.log
@@ -1,0 +1,7 @@
+
+Checking files for maximum 1000 lines...
+
+Checking Rust files in rust/...
+  Found 46 Rust file(s)
+
+Checked 46 file(s) - all within the line limit
