+++ rust/src/query_processor.rs
@@ -689,103 +689,6 @@ impl QueryProcessor {
         Ok(ResolvedLink::new(index, source, target, name))
     }
 
-    fn preserve_existing_substitution_parts(
-        storage: &mut impl NamedTypeLinks,
-        pattern: &Pattern,
-        solution: &mut HashMap<String, u32>,
-        index: u32,
-        source: &mut u32,
-        target: &mut u32,
-        visited_indexes: &mut HashSet<u32>,
-    ) -> Result<()> {
-        if !Self::is_normal_index(index) || !storage.exists(index) {
-            return Ok(());
-        }
-
-        if !visited_indexes.insert(index) {
-            return Ok(());
-        }
-
-        let existing_link = storage.get_link(index).ok_or(LinkError::NotFound(index))?;
-
-        if Self::should_preserve_existing_part(pattern.source.as_deref(), solution)
-            && Self::can_preserve_existing_part(
-                existing_link,
-                existing_link.source,
-                visited_indexes,
-            )
-        {
-            *source = existing_link.source;
-            Self::assign_variable_from_pattern(pattern.source.as_deref(), *source, solution);
-        } else if let Some(bound_source) =
-            Self::resolved_variable_part(pattern.source.as_deref(), solution)
-        {
-            *source = bound_source;
-        }
-
-        if Self::should_preserve_existing_part(pattern.target.as_deref(), solution)
-            && Self::can_preserve_existing_part(
-                existing_link,
-                existing_link.target,
-                visited_indexes,
-            )
-        {
-            *target = existing_link.target;
-            Self::assign_variable_from_pattern(pattern.target.as_deref(), *target, solution);
-        } else if let Some(bound_target) =
-            Self::resolved_variable_part(pattern.target.as_deref(), solution)
-        {
-            *target = bound_target;
-        }
-
-        visited_indexes.remove(&index);
-        Ok(())
-    }
-
-    fn should_preserve_existing_part(
-        pattern: Option<&Pattern>,
-        solution: &HashMap<String, u32>,
-    ) -> bool {
-        pattern.is_some_and(|pattern| {
-            pattern.is_leaf()
-                && Self::is_variable(&pattern.index)
-                && !solution.contains_key(&pattern.index)
-        })
-    }
-
-    fn resolved_variable_part(
-        pattern: Option<&Pattern>,
-        solution: &HashMap<String, u32>,
-    ) -> Option<u32> {
-        pattern.and_then(|pattern| {
-            if pattern.is_leaf() && Self::is_variable(&pattern.index) {
-                solution.get(&pattern.index).copied()
-            } else {
-                None
-            }
-        })
-    }
-
-    fn assign_variable_from_pattern(
-        pattern: Option<&Pattern>,
-        value: u32,
-        solution: &mut HashMap<String, u32>,
-    ) {
-        if let Some(pattern) = pattern {
-            Self::assign_variable(&pattern.index, value, solution);
-        }
-    }
-
-    fn can_preserve_existing_part(
-        existing_link: Link,
-        part: u32,
-        visited_indexes: &HashSet<u32>,
-    ) -> bool {
-        existing_link.is_full_point()
-            || existing_link.is_partial_point()
-            || !visited_indexes.contains(&part)
-    }
-
     fn resolve_identifier(
         &self,
         storage: &mut impl NamedTypeLinks,
